use nint with Unsafe.Add

+++ src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromCmykAvx2.cs
@@ -37,8 +37,8 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
                 // Used for the color conversion
                 var scale = Vector256.Create(1 / this.MaximumValue);
 
-                int n = values.Component0.Length / 8;
-                for (int i = 0; i < n; i++)
+                nint n = values.Component0.Length / 8;
+                for (nint i = 0; i < n; i++)
                 {
                     ref Vector256<float> c = ref Unsafe.Add(ref c0Base, i);
                     ref Vector256<float> m = ref Unsafe.Add(ref c1Base, i);
